requirements.txt
torch==2.8.0 --index-url https://download.pytorch.org/whl/cu126
torchvision==0.23.0

tqdm==4.67.1
numpy==2.3.3
scipy==1.16.3
scikit-learn==1.7.2
matplotlib==3.10.6
pillow==11.3.0
pandas==2.3.2

jupyter==1.1.1
notebook==7.4.7

albumentations==2.0.8
opencv-python>=4.10.0.82

seaborn==0.13.2
matplotlib==3.10.6
plotly==6.4.0